Added final reminder & restructured files for clarity
