+++ handlers/sessionendedrequesthandler.js
@@ -1,0 +1,13 @@
+const constants = require('./commonconstants')
+
+module.exports = {
+    canHandle(handlerInput) {
+      const request = handlerInput.requestEnvelope.request;
+      return request.type === 'SessionEndedRequest';
+    },
+    handle(handlerInput) {
+      console.log(`Session ended with reason: ${handlerInput.requestEnvelope.request.reason}`);
+      return handlerInput.responseBuilder
+        .getResponse();
+    },
+  };
