+++ handlers/starthandler.js
@@ -1,0 +1,8 @@
+const constants = require('./commonconstants')
+const {speechResponses, SKILL_NAME} = constants('../')
+
+module.exports = {
+    // retrieve data from DDB using session.user.userId and spoken name of longest ingredient
+    // call function to set timers
+    // remove data from DDB if successful
+}
