+++ lib/alarmhandler.js
@@ -42,9 +42,8 @@ const callAPI = (token, url, payload = null) => new Promise((resolve, reject) =>
 const SetAlarmHandler = (textContent, offsetSecs, endpoint, token) => {
     return new Promise((resolve, reject) => {
         let currentTime = new Date();
-        if (offsetSecs < 5) {
-            offsetSecs = 5
-        }
+        offsetSecs = (offsetSecs < 5) ? 5 : offsetSecs 
+        let reminderContent = (textContent === 'Complete') ? 'Everything should now be complete' : 'Time to start ' + textContent
         let useBody = {
             requestTime : currentTime.toISOString(),
             trigger: {
@@ -55,7 +54,7 @@ const SetAlarmHandler = (textContent, offsetSecs, endpoint, token) => {
                 spokenInfo: {
                     content: [{
                         locale: "en-US", 
-                        text: "Time to start " + textContent
+                        text: reminderContent
                     }]
                 }
             },
@@ -82,7 +81,14 @@ module.exports = (alarmList = [], endpoint, token) => {
         let alarms = []
         let maxAlarms = alarmList.length 
         let alarmsCompleted = 0
-    
+
+        // push 'all completed' entry onto alarmList array
+        alarmList.push({
+            name: 'Complete',
+            duration: 0,
+            startsecs: (alarmList[0].duration * 60)
+        })
+
         for (let alarm of alarmList) {
             SetAlarmHandler(alarm.name, alarm.startsecs, endpoint, token)
                 .then(alarmResult => {
