+++ lib/speechresponses.js
@@ -40,6 +40,7 @@ module.exports = (intentType) => {
       REQ_PLAYBACK_REMINDERS_1: "Ok. I will set {{var.0}} reminder{{var.1}} for you. ",
       REQ_PLAYBACK_REMINDERS_2a: "{{var.0}} immediately. ",
       REQ_PLAYBACK_REMINDERS_2b: "{{var.0}} in {{var.1}} minute{{var.2}}. ",
+      REQ_PLAYBACK_REMINDERS_2c: "and a final reminder in {{var.0}} minute{{var.1}}, when everything should be finished. ",
       REQ_PLAYBACK_REMINDERS_3: "Is that Ok? ",
       CFM_REMINDERS_SET: "Your reminders have been set. ",
       CFM_REMINDERS_FAILED: "Sorry, I was unable to set at least one of your reminders. ",
